Deprecate ::PrincessBride
