--- doc/unreleased/default/princess_bride.md
@@ -1,9 +1,0 @@
-# Faker::PrincessBride
-
-Available since version 1.9.0.
-
-```ruby
-Faker::PrincessBride.character #=> "Dread Pirate Roberts"
-
-Faker::PrincessBride.quote #=> "Hello. My name is Inigo Montoya. You killed my father. Prepare to die!"
-```
