--- lib/faker/default/princess_bride.rb
@@ -1,15 +1,0 @@
-# frozen_string_literal: true
-
-module Faker
-  class PrincessBride < Base
-    class << self
-      def character
-        fetch('princess_bride.characters')
-      end
-
-      def quote
-        fetch('princess_bride.quotes')
-      end
-    end
-  end
-end
